+++ container/Dockerfile.none
@@ -46,7 +46,7 @@ ARG CARGO_BUILD_JOBS
 
 ENV CARGO_TARGET_DIR=/workspace/target
 
-RUN cargo build --release --locked --features mistralrs,python && \
+RUN cargo build --release --locked && \
     cargo doc --no-deps && \
     cp target/release/dynamo-run /usr/local/bin && \
     cp target/release/http /usr/local/bin && \
