+++ launch/dynamo-run/src/opt.rs
@@ -107,13 +107,6 @@ pub enum Output {
     // Start vllm in a sub-process connecting via nats
     // Sugar for `python vllm_inc.py --endpoint <thing> --model <thing>`
     Vllm,
-
-    /// Run inference using a user supplied python file that accepts and returns
-    /// strings. It does it's own pre-processing.
-    #[cfg(feature = "python")]
-    PythonStr(String),
-    // DEVELOPER NOTE
-    // If you add an engine add it to `available_engines` below, and to Default if it makes sense
 }
 
 impl TryFrom<&str> for Output {
@@ -145,14 +138,6 @@ impl TryFrom<&str> for Output {
                 Ok(Output::Dynamic)
             }
 
-            #[cfg(feature = "python")]
-            python_str_gen if python_str_gen.starts_with(crate::PYTHON_STR_SCHEME) => {
-                let path = python_str_gen
-                    .strip_prefix(crate::PYTHON_STR_SCHEME)
-                    .unwrap();
-                Ok(Output::PythonStr(path.to_string()))
-            }
-
             e => Err(anyhow::anyhow!("Invalid out= option '{e}'")),
         }
     }
@@ -175,9 +160,6 @@ impl fmt::Display for Output {
             Output::EchoCore => "echo_core",
 
             Output::Dynamic => "dyn",
-
-            #[cfg(feature = "python")]
-            Output::PythonStr(_) => "pystr",
         };
         write!(f, "{s}")
     }
@@ -217,11 +199,6 @@ impl Output {
         out.push(Output::SgLang.to_string());
         out.push(Output::Trtllm.to_string());
         out.push(Output::Vllm.to_string());
-
-        #[cfg(feature = "python")]
-        {
-            out.push(Output::PythonStr("file.py".to_string()).to_string());
-        }
 
         out
     }
