+++ lib/bindings/python/Cargo.toml
@@ -40,8 +40,11 @@ block-manager = ["dynamo-llm/block-manager", "dep:dlpark"]
 [dependencies]
 dynamo-llm = { path = "../../llm" }
 dynamo-runtime = { path = "../../runtime" }
-dynamo-engine-python = { path = "../../engines/python" }
 
+anyhow = { version = "1" }
+async-stream = { version = "0.3" }
+async-trait = { version = "0.1" }
+async-openai = "0.27.2"
 futures = { version = "0.3" }
 once_cell = { version = "1.20.3" }
 serde = { version = "1" }
